MasterProject/requirements.txt
Django==5.0.1
django-environ==0.11.2
Pillow==10.2.0
pytesseract==0.3.10
gunicorn==21.2.0
whitenoise==6.6.0
tensorflow==2.8.0
protobuf==3.19.0
opencv-python==4.9.0.80
numpy==1.26.4
pipreqs==0.4.13
pymongo==4.6.2
celery[redis]